TrainingScheduler/requirements.txt
flask
flask_expects_json
pika
pandas
flask_sqlalchemy
psycopg2-binary
flask-cors